internal/pkg/scaffold: move some utility functions to internal/util/k8sutil
