+++ internal/pkg/scaffold/crd.go
@@ -169,7 +169,7 @@ func (s *CRD) CustomRender() ([]byte, error) {
 	}
 
 	setCRDVersions(crd)
-	return k8sutil.GetObjectBytes(crd)
+	return k8sutil.GetObjectBytes(crd, yaml.Marshal)
 }
 
 func newCRDForResource(r *Resource) *apiextv1beta1.CustomResourceDefinition {
