+++ internal/pkg/scaffold/olm-catalog/csv_test.go
@@ -167,33 +167,6 @@ func TestUpdateVersion(t *testing.T) {
 	}
 }
 
-func TestGetDisplayName(t *testing.T) {
-	cases := []struct {
-		input, wanted string
-	}{
-		{"Appoperator", "Appoperator"},
-		{"appoperator", "Appoperator"},
-		{"appoperatoR", "Appoperato R"},
-		{"AppOperator", "App Operator"},
-		{"appOperator", "App Operator"},
-		{"app-operator", "App Operator"},
-		{"app-_operator", "App Operator"},
-		{"App-operator", "App Operator"},
-		{"app-_Operator", "App Operator"},
-		{"app--Operator", "App Operator"},
-		{"app--_Operator", "App Operator"},
-		{"APP", "APP"},
-		{"another-AppOperator_againTwiceThrice More", "Another App Operator Again Twice Thrice More"},
-	}
-
-	for _, c := range cases {
-		dn := getDisplayName(c.input)
-		if dn != c.wanted {
-			t.Errorf("Wanted %s, got %s", c.wanted, dn)
-		}
-	}
-}
-
 func TestSetAndCheckOLMNamespaces(t *testing.T) {
 	depBytes, err := ioutil.ReadFile(filepath.Join(testDeployDir, "operator.yaml"))
 	if err != nil {
